Complex mysql queries are added and compile time polymorphism code is written
