+++ ArrayList/bin/.gitignore
@@ -1,0 +1,4 @@
+/String/
+/com/
+/components/
+/org/
