+++ Graph/bin/.gitignore
@@ -1,3 +1,2 @@
 /module-info.class
 /GraphNode.class
-/org/
