+++ Tree_Lucky_Kumar/bin/.gitignore
@@ -1,0 +1,1 @@
+/org/
